Create CSRF token values that are valid Websocket subprotocol names
WARNING!

    The way we generate the CSRF tokens in this example (by using a
    fixed value) *IS NOT SECURE*. Do *NOT* use this example as the
    basis for production code.

    This example is only meant as an easy way to test the proposed
    feature named "Send CSRF tokens in the Sec-WebSocket-Protocol
    header".

    CSRF tokens should be un-guessable by third-parties, and unique for
    each user session or each user request (which is not the case in this
    example).

END OF WARNING.

As we are sending the CSRF token value in the Sec-WebSocket-Protocol
request header, we need to make sure that the values we send in there
are accepted by both the browser (the websocket client side) and the
HTTP server.

RFC-6455, section 4.3 (Collected ABNF for New Header Fields Used in
Handshake), lists the ABNF grammar rules for the 'Sec-WebSocket-Protocol'
header, both for the client requests (Sec-WebSocket-Protocol-Client) and
the server responses (Sec-WebSocket-Protocol-Server). They are defined
as:

    Sec-WebSocket-Protocol-Client = 1#token
    Sec-WebSocket-Protocol-Server = token

And the same section states that the ABNF grammar rules for 'token' is
defined in RFC-2616.

RFC-2616, section 2.2 (Basic Rules), lists the ABNF grammar rules for
'token' as:

       token          = 1*<any CHAR except CTLs or separators>
       separators     = "(" | ")" | "<" | ">" | "@"
                      | "," | ";" | ":" | "\" | <">
                      | "/" | "[" | "]" | "?" | "="
                      | "{" | "}" | SP | HT
       CTL            = <any US-ASCII control character
                        (octets 0 - 31) and DEL (127)>

Unfortunately, the anti-forgery tokens generated by the
ring.middleware.anti-forgery library contain some of the values in the
'separators' grammar rule. Because it generates anti-forgery tokens as
random vales, encoded as "normal"[1] Base64 encoded strings. And those
that can include the '/' character in them, which is part of the
'separators' rule, and thus an invalid character for a 'token'. Thus
the generated values are unsuitable for the Sec-WebSocket-Protocol
header value.

That is why we generate our own CSRF token values that comply with the
ABNF grammar rules for 'token'. Otherwise, web browsers such as
Firefox throw an exception when trying to create a WebSocket with an
invalid subprotocol name.

[1] See RFC-4868, section 4 ("Base 64 Encoding").